Novuslogproject/requirements.txt
anyio==4.0.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.4
braintree==4.26.0
Brlapi==0.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
comm==0.1.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dblatex===0.3.11py3
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.6
django-cors-headers==4.3.1
django-import-export==3.3.6
django-jazzmin==2.6.0
djangorestframework==3.14.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
Faker==20.0.3
fasteners==0.14.1
fastjsonschema==2.18.1
filelock==3.0.12
fqdn==1.5.1
future==0.18.2
google-ai-generativelanguage==0.1.0
google-api-core==2.15.0
google-auth==2.26.1
google-generativeai==0.1.0rc1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httplib2==0.14.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.5
more-itertools==4.2.0
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.10.4
notebook==7.0.6
notebook_shim==0.2.3
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
outcome==1.3.0
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.16.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
python-json-logger==2.0.7
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.9
SecretStorage==2.3.1
selenium==4.14.0
Send2Trash==1.8.2
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
tablib==3.5.0
terminado==0.17.1
tinycss2==1.2.1
tkintertable==1.3.3
tomli==2.0.1
tornado==6.3.3
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.0.7
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wsproto==1.2.0
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0
